fix up compile errors and imports within the reportserver tests
