+++ tests/reportserver/README.md
@@ -1,1 +1,1 @@
-holds the testing code
+Holds the testing code for the reportserver
